Move some examples to snippets
