+++ wpilibjExamples/src/main/java/org/wpilib/snippets/addressableled/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.addressableled;
+package org.wpilib.snippets.addressableled;
 
 import static org.wpilib.units.Units.Meters;
 import static org.wpilib.units.Units.MetersPerSecond;
