+++ wpilibjExamples/src/main/java/org/wpilib/snippets/apriltagsvision/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.apriltagsvision;
+package org.wpilib.snippets.apriltagsvision;
 
 import java.util.ArrayList;
 import org.opencv.core.Mat;
