+++ wpilibjExamples/src/main/java/org/wpilib/snippets/canpdp/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.canpdp;
+package org.wpilib.snippets.canpdp;
 
 import org.wpilib.framework.TimedRobot;
 import org.wpilib.hardware.power.PowerDistribution;
