+++ wpilibjExamples/src/main/java/org/wpilib/snippets/digitalcommunication/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.digitalcommunication;
+package org.wpilib.snippets.digitalcommunication;
 
 import java.util.Optional;
 import org.wpilib.driverstation.DriverStation;
