+++ wpilibjExamples/src/main/java/org/wpilib/snippets/dutycycleinput/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.dutycycleinput;
+package org.wpilib.snippets.dutycycleinput;
 
 import org.wpilib.framework.TimedRobot;
 import org.wpilib.hardware.rotation.DutyCycle;
