+++ wpilibjExamples/src/main/java/org/wpilib/snippets/i2ccommunication/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.i2ccommunication;
+package org.wpilib.snippets.i2ccommunication;
 
 import java.util.Optional;
 import org.wpilib.driverstation.DriverStation;
