+++ wpilibjExamples/src/main/java/org/wpilib/snippets/snippets.json
@@ -171,5 +171,155 @@
     "foldername": "profiledpidfeedforward",
     "gradlebase": "java",
     "robotclass": "Robot"
+  },
+  {
+    "name": "Addressable LED",
+    "description": "Display a rainbow pattern on an addressable LED strip.",
+    "tags": [
+      "Hardware",
+      "Basic Robot",
+      "AddressableLEDs"
+    ],
+    "foldername": "addressableled",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "AprilTags Vision",
+    "description": "On-roboRIO detection of AprilTags using an attached USB camera.",
+    "tags": [
+      "Vision",
+      "AprilTags"
+    ],
+    "foldername": "apriltagsvision",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "PDP CAN Monitoring",
+    "description": "Monitor Power Distribution data such as voltage, current, temperature, etc.",
+    "tags": [
+      "Hardware",
+      "PDP",
+      "SmartDashboard"
+    ],
+    "foldername": "canpdp",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "Digital Communication Sample",
+    "description": "Communicates with external devices (such as an Arduino) using the roboRIO's DIO.",
+    "tags": [
+      "Hardware",
+      "Digital Output"
+    ],
+    "foldername": "digitalcommunication",
+    "gradlebase": "java",
+    "robotclass": "Robot",
+    "hasunittests": true
+  },
+  {
+    "name": "Duty Cycle Input",
+    "description": "View duty-cycle input.",
+    "tags": [
+      "Hardware",
+      "Duty Cycle",
+      "SmartDashboard"
+    ],
+    "foldername": "dutycycleinput",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "EventLoop",
+    "description": "Manage a ball system using EventLoop and BooleanEvent.",
+    "tags": [
+      "Basic Robot",
+      "Flywheel",
+      "EventLoop"
+    ],
+    "foldername": "eventloop",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "Flywheel BangBangController",
+    "description": "A sample program to demonstrate the use of a BangBangController with a flywheel to control RPM",
+    "tags": [
+      "Flywheel",
+      "Simulation",
+      "Joystick"
+    ],
+    "foldername": "flywheelbangbangcontroller",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "HTTP Camera",
+    "description": "Acquire images from an HTTP network camera and adds some annotation to the image (as you might do for showing operators the result of some image recognition), and sends it to the dashboard for display.",
+    "tags": [
+      "Vision"
+    ],
+    "foldername": "httpcamera",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "I2C Communication Sample",
+    "description": "Communicate with external devices (such as an Arduino) using the roboRIO's I2C port.",
+    "tags": [
+      "Hardware",
+      "I2C"
+    ],
+    "foldername": "i2ccommunication",
+    "gradlebase": "java",
+    "robotclass": "Robot",
+    "hasunittests": true
+  },
+  {
+    "name": "Intermediate Vision",
+    "description": "Acquire images from an attached USB camera and add some annotation to the image (as you might do for showing operators the result of some image recognition) and send it to the dashboard for display.",
+    "tags": [
+      "Vision"
+    ],
+    "foldername": "intermediatevision",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "Motor Control",
+    "description": "Control a single motor with a joystick, displaying the movement of the motor using an encoder.",
+    "tags": [
+      "Basic Robot",
+      "Encoder",
+      "SmartDashboard",
+      "Joystick"
+    ],
+    "foldername": "motorcontrol",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "Simple Vision",
+    "description": "Use the CameraServer class to stream from a USB Webcam without processing the images.",
+    "tags": [
+      "Vision"
+    ],
+    "foldername": "quickvision",
+    "gradlebase": "java",
+    "robotclass": "Robot"
+  },
+  {
+    "name": "Solenoids",
+    "description": "Control a single and double solenoid from joystick buttons.",
+    "tags": [
+      "Hardware",
+      "Joystick",
+      "Pneumatics"
+    ],
+    "foldername": "solenoid",
+    "gradlebase": "java",
+    "robotclass": "Robot"
   }
 ]
