+++ wpilibjExamples/src/main/java/org/wpilib/snippets/solenoid/Robot.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.solenoid;
+package org.wpilib.snippets.solenoid;
 
 import org.wpilib.driverstation.Joystick;
 import org.wpilib.framework.TimedRobot;
