+++ wpilibjExamples/src/test/java/org/wpilib/snippets/digitalcommunication/DigitalCommunicationTest.java
@@ -2,36 +2,33 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package org.wpilib.examples.i2ccommunication;
+package org.wpilib.snippets.digitalcommunication;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertTimeoutPreemptively;
+import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.time.Duration;
-import java.util.concurrent.CompletableFuture;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.parallel.ResourceLock;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EnumSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import org.wpilib.driverstation.DriverStation;
 import org.wpilib.hardware.hal.AllianceStationID;
 import org.wpilib.hardware.hal.HAL;
 import org.wpilib.hardware.hal.RobotMode;
-import org.wpilib.simulation.CallbackStore;
+import org.wpilib.simulation.DIOSim;
 import org.wpilib.simulation.DriverStationSim;
-import org.wpilib.simulation.I2CSim;
 import org.wpilib.simulation.SimHooks;
 
 @ResourceLock("timing")
-class I2CCommunicationTest {
+class DigitalCommunicationTest {
   private Robot m_robot;
   private Thread m_thread;
-  private final I2CSim m_i2c = new I2CSim(Robot.kPort.value);
-  private CompletableFuture<String> m_future;
-  private CallbackStore m_callback;
+  private final DIOSim m_allianceOutput = new DIOSim(Robot.kAlliancePort);
+  private final DIOSim m_enabledOutput = new DIOSim(Robot.kEnabledPort);
+  private final DIOSim m_autonomousOutput = new DIOSim(Robot.kAutonomousPort);
+  private final DIOSim m_alertOutput = new DIOSim(Robot.kAlertPort);
 
   @BeforeEach
   void startThread() {
@@ -39,10 +36,6 @@ class I2CCommunicationTest {
     SimHooks.pauseTiming();
     SimHooks.setProgramStarted(false);
     DriverStationSim.resetData();
-    m_future = new CompletableFuture<>();
-    m_callback =
-        m_i2c.registerWriteCallback(
-            (name, buffer, count) -> m_future.complete(new String(buffer, 0, count)));
     m_robot = new Robot();
     m_thread = new Thread(m_robot::startCompetition);
     m_thread.start();
@@ -59,8 +52,10 @@ class I2CCommunicationTest {
       Thread.currentThread().interrupt();
     }
     m_robot.close();
-    m_callback.close();
-    m_i2c.resetData();
+    m_allianceOutput.resetData();
+    m_enabledOutput.resetData();
+    m_autonomousOutput.resetData();
+    m_alertOutput.resetData();
   }
 
   @EnumSource(AllianceStationID.class)
@@ -69,17 +64,12 @@ class I2CCommunicationTest {
     DriverStationSim.setAllianceStationId(alliance);
     DriverStationSim.notifyNewData();
 
-    assertTrue(m_i2c.getInitialized());
+    assertTrue(m_allianceOutput.getInitialized());
+    assertFalse(m_allianceOutput.getIsInput());
 
     SimHooks.stepTiming(0.02);
 
-    String str = assertTimeoutPreemptively(Duration.ofMillis(20L), () -> m_future.get());
-    char expected = alliance.name().startsWith("Red") ? 'R' : 'B';
-    if (alliance.name().startsWith("Unknown")) {
-      expected = 'U';
-    }
-
-    assertEquals(expected, str.charAt(0));
+    assertEquals(alliance.name().startsWith("Red"), m_allianceOutput.getValue());
   }
 
   @ValueSource(booleans = {true, false})
@@ -88,14 +78,12 @@ class I2CCommunicationTest {
     DriverStationSim.setEnabled(enabled);
     DriverStationSim.notifyNewData();
 
-    assertTrue(m_i2c.getInitialized());
+    assertTrue(m_enabledOutput.getInitialized());
+    assertFalse(m_enabledOutput.getIsInput());
 
     SimHooks.stepTiming(0.02);
 
-    String str = assertTimeoutPreemptively(Duration.ofMillis(20L), () -> m_future.get());
-    char expected = enabled ? 'E' : 'D';
-
-    assertEquals(expected, str.charAt(1));
+    assertEquals(enabled, m_enabledOutput.getValue());
   }
 
   @ValueSource(booleans = {true, false})
@@ -104,28 +92,25 @@ class I2CCommunicationTest {
     DriverStationSim.setRobotMode(autonomous ? RobotMode.AUTONOMOUS : RobotMode.TELEOPERATED);
     DriverStationSim.notifyNewData();
 
-    assertTrue(m_i2c.getInitialized());
+    assertTrue(m_autonomousOutput.getInitialized());
+    assertFalse(m_autonomousOutput.getIsInput());
 
     SimHooks.stepTiming(0.02);
 
-    String str = assertTimeoutPreemptively(Duration.ofMillis(20L), () -> m_future.get());
-    char expected = autonomous ? 'A' : 'T';
-
-    assertEquals(expected, str.charAt(2));
+    assertEquals(autonomous, m_autonomousOutput.getValue());
   }
 
-  @ValueSource(doubles = {112.0, 45.0, 27.0, 23.0, 3.0})
-  @ParameterizedTest(name = "matchTime[{index}]: {0}s")
-  void matchTimeTest(double matchTime) {
+  @ValueSource(doubles = {45.0, 27.0, 23.0})
+  @ParameterizedTest(name = "alert[{index}]: {0}s")
+  void alertTest(double matchTime) {
     DriverStationSim.setMatchTime(matchTime);
     DriverStationSim.notifyNewData();
-    assertTrue(m_i2c.getInitialized());
+
+    assertTrue(m_alertOutput.getInitialized());
+    assertFalse(m_alertOutput.getIsInput());
 
     SimHooks.stepTiming(0.02);
 
-    String str = assertTimeoutPreemptively(Duration.ofMillis(20L), () -> m_future.get());
-    String expected = String.format("%03d", (int) DriverStation.getMatchTime());
-
-    assertEquals(expected, str.substring(3));
+    assertEquals(matchTime <= 30 && matchTime >= 25, m_alertOutput.getValue());
   }
 }
